api/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
django-restframework==0.0.1
sqlparse==0.5.3
